Jackrabbit bug affecting transaction locking
I found a bug in Jackrabbit which is affecting transaction locking.
I'm not versed with the architecture of Jackrabbit but this is what I found out:

A Jackrabbit user has to present lock tokens to mark the beginning and the end (and every request via davex?) of a transaction.
I guess the lock token is based on an uuid with an additional digit at the end which is some kind of check digit.
But Jackrabbit just gives away lock tokens without the check digit but tries to check the digit again while managing the tokens.

So my idea was to forget about the check digit. I don't know if this is a good idea. ;-) And I don't know if I caught every point in the code where a change would be necessary. Jackrabbit seems to run properly.